Dev: move getProviderFromChainId to sdk
